Time: 0 ms (100.00%), Space: 43.3 MB (82.15%) - LeetHub

+++ 98-validate-binary-search-tree/98-validate-binary-search-tree.java
@@ -21,9 +21,8 @@ class Solution {
     public boolean isValid(long lowerLimit, long upperLimit, TreeNode root) {
         if (root==null) return true;
         if (lowerLimit<root.val && upperLimit>root.val) {
-            return isValid(lowerLimit, (long)root.val, root.left) && isValid((long)root.val, upperLimit, root.right);
+            return isValid(lowerLimit, root.val, root.left) && isValid(root.val, upperLimit, root.right);
         } 
         return false;
     }
-    
 }
